Nautical Miles (the updated version)

+++ src/apcs/project1/Project3.java
@@ -4,8 +4,8 @@ public class Project3 {
     public static void main(String[] args) {
         Scanner reader = new Scanner(System.in);
         System.out.println("Enter number of Kilometers: ");
-        int x = reader.nextInt();
-        int y = ((x * 60 * 90)/10000);
+        double x = reader.nextInt();
+        double y = ((x * 60 * 90)/10000);
         System.out.println("# of Nautical Miles: " + y);
     }
 }
